requirements.txt
azure-functions
pandas==0.24.2
numpy==1.15.4
xgboost==0.90
flask